code review changes
* renamed resource file info provider to file context provider
* removed unnecessary work in file context provider 
* renamed interchange pipeline to file import pipeline
