+++ Utilities/DataLoading/EdFi.LoadTools/BulkLoadClient/LoadProcess.cs
@@ -14,7 +14,7 @@ using EdFi.LoadTools.ApiClient;
 using EdFi.LoadTools.BulkLoadClient.Application;
 using EdFi.LoadTools.Engine;
 using EdFi.LoadTools.Engine.Factories;
-using EdFi.LoadTools.Engine.InterchangePipeline;
+using EdFi.LoadTools.Engine.FileImportPipeline;
 using EdFi.LoadTools.Engine.Mapping;
 using EdFi.LoadTools.Engine.MappingFactories;
 using EdFi.LoadTools.Engine.ResourcePipeline;
@@ -145,7 +145,7 @@ namespace EdFi.LoadTools.BulkLoadClient
             container.RegisterSingleton<SwaggerMetadataRetriever>();
             container.RegisterSingleton<SwaggerRetriever>();
             container.RegisterSingleton<XsdStreamsRetriever>();
-            container.RegisterSingleton<IResourceToFileTypeInfoProvider, ResourceToFileTypeInfoProvider>();
+            container.RegisterSingleton<IFileContextProvider, FileContextProvider>();
 
             container.RegisterSingleton<IEnumerable<JsonModelMetadata>>(
                 () => container.GetInstance<IMetadataFactory<JsonModelMetadata>>().GetMetadata().ToArray());
@@ -168,7 +168,7 @@ namespace EdFi.LoadTools.BulkLoadClient
             container.RegisterSingleton<IMetadataMappingFactory, ResourceToResourceMetadataMappingFactory>();
             container.RegisterSingleton<ITokenRetriever, TokenRetriever>();
 
-            container.RegisterCollection<IInterchangePipelineStep>(
+            container.RegisterCollection<IFileImportPipelineStep>(
                 new[]
                 {
                     typeof(FindReferencesStep), typeof(PreloadReferencesStep)
